+++ core/jvm/src/main/kotlin/com/github/jcornaz/miop/experimental/Operators.kt
@@ -130,7 +130,12 @@ public fun <E> ReceiveChannel<E>.launchConsumeEach(
     consumeEach { action(it) }
 }
 
-internal fun <E> ReceiveChannel<E>.distinctUntilChanged(): ReceiveChannel<E> = produce(Unconfined) {
+/**
+ * Returns a [ReceiveChannel] which emits the element of this channel, unless the element is equal to the last emitted element.
+ *
+ * Example: for the source: [1, 2, 2, 1, 2] [distinctUntilChanged] will emit: [1, 2, 1, 2]
+ */
+public fun <E> ReceiveChannel<E>.distinctUntilChanged(): ReceiveChannel<E> = produce(Unconfined) {
     consume {
         var latest = receive()
         send(latest)